refactor: merge code from wrapper into executor (match original structure)
